testing to see what checking out single file does

+++ docs/pythonfullinstall.sh
@@ -1,6 +1,10 @@
-
+#test line - MAVARA
 USEOTHER=0
-USEKRAKEN=1
+USEKRAKEN=0
+USEPFE=1
+
+alias cp='cp'
+alias mv='mv'
 
 # see:
 #http://docs.python.org/install/index.html
@@ -16,8 +20,8 @@ fi
 if [ $USEKRAKEN -eq 1 ]
 then
     # For Kraken:
-    SRCDIR=/lustre/scratch/jmckinne/tarballs
-    BASE=/lustre/scratch/jmckinne/
+    SRCDIR=/lustre/scratch/$USER/tarballs
+    BASE=/lustre/scratch/$USER/
 
     #source ~/setuppython27
     # trying to get MKL to work
@@ -26,32 +30,78 @@ then
     module unload python/2.7.1-cnl
 fi
 
+if [ $USEPFE -eq 1 ]
+then
+    # For PFE:
+    SRCDIR=/nobackup/$USER/tarballs/
+    #BASE=/u/$USER/
+    BASE=/nobackup/$USER/
+
+fi
+
+
 
 # make tools path as consistent with ~/.bashrc line:
 mkdir -p $BASE/lib/python
 
+# get tarballs, e.g.:
+# cd $SCRDIR
+# scp -rp jmckinne@ki-jmck.slac.stanford.edu:/data1/jmckinne/pythonstuff/tarballs .
 
 # PYTHON
 cd $SRCDIR
 #tar xvzf Python-2.7.2.tgz
 cd Python-2.7.2/
 make clean
-if [ $USEKRAKEN ]
+# on normal system using icc
+#./configure --prefix=$BASE/
+if [ $USEKRAKEN -eq 1 ]
 then
     # worse:
     #export CC=cc
+    # on normal system using icc
+    ./configure --prefix=$BASE/
 fi
+if [ $USEPFE -eq 1 ]
+then
+#mkdir ~/bin/
+#cd ~/bin/
+#ln -s /nasa/intel/Compiler/2012.0.032/bin/icc cc
+#export PATH=$HOME/bin/$PATH  # and add this to ~/.profile
+## on PFE:
+#cd $SRCDIR
+#cd Python-2.7.2/
+#LDFLAGS=/nasa/intel/Compiler/2012.0.032/mkl/lib/intel64/ -lmkl -lmkl_scalapack_lp64
+#CPPFLAGS=-I/nasa/intel/Compiler/2012.0.032/mkl/include/
+#./configure --without-gcc --prefix=$BASE/
+# problem with MKL on PFE, so just stick with gcc and lapack/blas from netlib.
+LDFLAGS=
+CPPFLAGS=
+rm -rf ~/bin/cc
 ./configure --prefix=$BASE/
+fi
+
 make
+# ok that curses doesn't work, but no other packages should fail.
 make install
 # then logout-login to ensure python path set right because still loads old python even if which python points to new!
+#       or at least check that new python really loads by loading it and checking what reversion it reports.
 
+# add below exports to bash start-up script (maybe .bashrc or .profile)
 export PYTHONPATH=$BASE/lib/python/:$BASE/py/
-export PATH=$BASE/bin:$PATH
+export PATH=$HOME/bin/:$BASE/bin:$PATH
 export PYTHON_LIB=$BASE/lib/
 export PYTHON_INC=$BASE/include/python2.7/
-export LD_LIBRARY_PATH=$BASE/lib/:$LD_LIBRARY_PATH
+export LD_LIBRARY_PATH=$HOME/lib/:$BASE/lib/:$LD_LIBRARY_PATH
 export LIBRARY_PATH=$BASE/lib/:$LIBRARY_PATH
+
+if [ $USEPFE -eq 1 ]
+then
+cd $BASE/lib
+ln -s python2.7 python
+#mv $BASE/lib/python2.7/site.py $BASE/lib/python2.7/site.py.backup
+mv /nobackupp8/jmckinn2/lib/python2.7/site.py /nobackupp8/jmckinn2/lib/asdf.py
+fi
 
 cd $SRCDIR
 #tar xvzf setuptools-0.6c11.tar.gz
@@ -59,12 +109,18 @@ cd setuptools-0.6c11/
 rm -rf build
 python setup.py install --home=$BASE
 
+# don't do below.
+# mv $BASE/lib/python2.7/site.py.backup $BASE/lib/python2.7/site.py
+
+# doesn't work on PFE for some reason
+if [ $USEPFE -eq 0 ]
+then
 cd $SRCDIR
 #tar xvzf python-dateutil-1.5.tar.gz
 cd python-dateutil-1.5/
 rm -rf build
 python setup.py install --home=$BASE
-
+fi
 
 cd $SRCDIR
 #tar xvzf yasm-1.1.0.tar.gz
@@ -98,6 +154,36 @@ rm -rf build
 python setup.py install --home=$BASE
 
 
+
+
+if [ $USEPFE -eq 1 ]
+then
+cd $SRCDIR
+wget http://www.netlib.org/blas/blas.tgz
+cd BLAS
+make
+mkdir ~/lib/
+cp blas_LINUX.a ~/lib/
+#
+cd $SRCDIR
+wget http://www.netlib.org/lapack/lapack-3.5.0.tgz
+tar xvzf lapack-3.5.0.tgz
+cd lapack-3.5.0
+cp $SRCDIR/BLAS/blas_LINUX.a librefblas.a
+cp librefblas.a ~/lib/
+cp make.inc.example make.inc
+make
+cp liblapack.a libtmglib.a ~/lib/
+fi
+
+
+
+
+if [ $USEPFE -eq 0 ]
+then
+############
+# NON-PFE systems MKL
+
 # Intel Studio if not already installed
 cd /data/jon/
 # tar xvzf c_studio_xe_2011_update2.tgz
@@ -126,12 +212,18 @@ http://www.scipy.org/Installing_SciPy/Linux
 # Intel link advisor:
 # http://software.intel.com/en-us/articles/intel-mkl-link-line-advisor/
 
+
+# DONE WITH MKL
+###############################
+fi
+
 cd $SRCDIR
 #tar xvzf numpy-1.6.1.tar.gz
 cd numpy-1.6.1/
 rm -rf build
 # cp site.cfg.example site.cfg
 
+# Take [mkl] section, copy to site.cfg, and then uncomment that section.
 # For INTEL icc:
 # with MKL just do:
 # http://www.shocksolution.com/2008/09/installing-numpy-with-the-intel-math-kernel-library-mkl/
@@ -149,14 +241,34 @@ rm -rf build
 #include_dirs = /opt/intel/composer_xe_2011_sp1.7.256/mkl/include
 #lapack_libs = mkl_scalapack_lp64
 #mkl_libs = 
-# for Kraken, uses libsci instead of mkl
 
+
+# for Kraken, uses libsci instead of mkl
 # [mkl]
 # library_dirs = /opt/intel/composer_xe_2011_sp1.8.273/mkl/lib/intel64/
 # include_dirs = /opt/intel/composer_xe_2011_sp1.8.273/mkl/include/
 # lapack_libs = mkl_scalapack_lp64
 # mkl_libs =
 
+
+# for PFE with MKL (not working)
+
+# [mkl]
+# library_dirs = /nasa/intel/Compiler/2012.0.032/mkl/lib/intel64/
+# include_dirs = /nasa/intel/Compiler/2012.0.032/mkl/include/
+# lapack_libs = mkl_scalapack_lp64
+# mkl_libs =
+
+
+# PFE without MKL
+#[DEFAULT]
+# REPLACE jmckinn2 by your $USER
+#library_dirs = /usr/local/lib,/u/jmckinn2/lib
+#include_dirs = /usr/local/include,/u/jmckinn2/include
+#libraries = f77blas, cblas, atlas
+#[lapack_opt]
+#libraries = lapack, f77blas, cblas, atlas
+#              
 
 #
 
@@ -166,15 +278,25 @@ rm -rf build
 #     cc_exe = 'icc -O2 -g -openmp -fPIC  -L/opt/intel/composer_xe_2011_sp1.8.273/mkl/lib/intel64/ -lmkl -axv'
 #     cc_args = "-fPIC -DMKL_ILP64 "
 #
+# for Kraken, replace -L with what used above in [mkl] section.  Same for PFE.
+# for PFE, remove -axv
+#
+if [ $USEKRAKEN -eq 1 ]
+then
 python setup.py config --compiler=intelem build_clib --compiler=intelem build_ext --compiler=intelem install --home=$BASE
+fi
 
 ##########
 ## To use GCC:
+if [ $USEPFE -eq 1 ]
+then
 #python setup.py install --home=$BASE
+python setup.py config --compiler=unix --fcompiler=gnu95 install --home=$BASE
+fi
 
 # test numpy to see if anything really works (even compile):
 cd $SRCDIR
-# NOT WORKING for Kraken
+# NOT WORKING for Kraken or PFE
 python -c 'import numpy; numpy.test()'
 
 
@@ -191,9 +313,29 @@ cd $SRCDIR
 #tar xvzf python-scipy_0.9.0+dfsg1.orig.tar.gz
 cd scipy-0.9.0.orig/
 rm -rf build
+if [ $USEPFE -eq 0 ]
+then
 python setup.py install --home=$BASE
+fi
+
+if [ $USEPFE -eq 1 ]
+then
+export LAPACK=$HOME/lib/liblapack
+export LAPACK_SRC=$SRCDIR/lapack-3.5.0/
+export BLAS=$HOME/lib/blas_LINUX.a
+export BLAS_SRC=$SRCDIR/BLAS/
+cp $SRCDIR/lapack-3.5.0/liblapack.a $SRCDIR/lapack-3.5.0/libtmglib.a .
+cp $SRCDIR/BLAS/blas_LINUX.a librefblas.a
+cp $SRCDIR/BLAS/blas_LINUX.a libblas.a
+python setup.py config --compiler=unix --fcompiler=gnu95 install --home=$BASE
+fi
+
+#['setup.py', 'config', '--compiler=gcc', 'config_fc', '--fcompiler=gfortran', 'build_clib', '--compiler=gcc', '--fcompiler=gfortran', 'build_ext', '--compiler=gcc', '--fcompiler=gfortran', 'install', '--prefix=/nobackup/jmckinn2/']
 
 
+# DON'T USE NEW SCIPY
+if [ 1 -eq 0 ]
+then
 ####################
 # NEW scipy
 # http://www.scipy.org/Download
@@ -202,8 +344,8 @@ cd $SRCDIR
 cd scipy-0.10.0
 rm -rf build
 python setup.py config --compiler=intelem --fcompiler=intelem build_clib --compiler=intelem --fcompiler=intelem build_ext --compiler=intelem --fcompiler=intelem install --home=$BASE
-
 ##########
+fi
 
 
 
@@ -212,8 +354,17 @@ cd $SRCDIR
 # tar xvzf matplotlib.tgz
 cd matplotlib/matplotlib
 rm -rf build
+if [ $USEPFE -eq 1 ]
+then
+python setup.py config --compiler=unix --fcompiler=gnu95 install --home=$BASE
+fi
+if [ $USEPFE -eq 0 ]
+then
 python setup.py install --home=$BASE
+fi
 
+if [ $USEPFE -eq 0 ] # pfe already has it as a module one can load
+then
 # ffmpeg
 cd $SRCDIR
 #tar xvzf ffmpeg-0.8.tar.gz
@@ -224,7 +375,7 @@ make
 make install
 make libinstall
 make libainstall
-
+fi
 
 # try optimize.leastsq():
 # http://www.scipy.org/scipy_Example_List#head-4c436ae0085d9a56056425d11abff4ccdd3d3620
